return None if file not exist

+++ monty/shutil.py
@@ -100,15 +100,13 @@ def decompress_file(filepath) -> str | None:
 
     Args:
         filepath (str): Path to file.
-        compression (str): A compression mode. Valid options are "gz" or
-            "bz2". Defaults to "gz".
 
     Returns:
         str: The decompressed file path.
     """
     toks = filepath.split(".")
     file_ext = toks[-1].upper()
-    if file_ext in ["BZ2", "GZ", "Z"]:
+    if file_ext in ["BZ2", "GZ", "Z"] and os.path.isfile(filepath):
         decompressed_file = ".".join(toks[0:-1])
         with zopen(filepath, "rb") as f_in, open(decompressed_file, "wb") as f_out:
             f_out.writelines(f_in)
